Refactor : AccessToken 및 userInfo 로직 변경
