+++ src/components/@shared/provider/AccessTokenLoader.tsx
@@ -2,13 +2,9 @@
 
 import { useEffect } from 'react';
 import { useUserStore } from '@/store/userStore';
-import { useQuery } from '@tanstack/react-query';
-import { GetUser } from '@/axios/user';
 
 export default function AccessTokenLoader() {
   const setAccessToken = useUserStore((state) => state.setAccessToken);
-  const setUserInfo = useUserStore((state) => state.setUserInfo);
-  const accessToken = useUserStore((state) => state.accessToken);
 
   useEffect(() => {
     const token = localStorage.getItem('accessToken');
@@ -17,18 +13,5 @@ export default function AccessTokenLoader() {
     }
   }, [setAccessToken]);
 
-  const { data } = useQuery({
-    queryKey: ['userInfo'],
-    queryFn: () => GetUser(accessToken!),
-    enabled: !!accessToken,
-    retry: false,
-    staleTime: 1000 * 60 * 5,
-  });
-
-  useEffect(() => {
-    if (data) {
-      setUserInfo(data.data);
-    }
-  }, [data, setUserInfo]);
   return null;
 }
