--- src/hooks/useSocialLogin.ts
@@ -1,22 +1,0 @@
-import { useRouter } from 'next/navigation';
-import { API_PATH } from '@/axios/path.config';
-
-export const useSocialLogin = () => {
-  const router = useRouter();
-  const login = (social: string) => {
-    const url = API_PATH.auth.signUp(social);
-    window.open(url, '_blank', 'width=1024,height=768');
-
-    const handleMessage = (e: MessageEvent) => {
-      if (e.origin !== window.location.origin) return;
-
-      if (e.data.success && e.data.accessToken) {
-        localStorage.setItem('accessToken', e.data.accessToken);
-        router.push('/');
-        window.removeEventListener('message', handleMessage);
-      }
-    };
-    window.addEventListener('message', handleMessage);
-  };
-  return { login };
-};
